feat(admin): 添加用户管相关功能
1. 添加用户登录、注册功能
2. 添加用户管理相关功能
3. 添加用户-角色功能

+++ yooj-backend-admin/admin-boot/src/main/java/com/yocy/admin/model/entity/SysUser.java
@@ -17,7 +17,7 @@ public class SysUser extends BaseEntity {
      * 
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 用户名
@@ -42,7 +42,7 @@ public class SysUser extends BaseEntity {
     /**
      * 部门ID
      */
-    private Integer deptId;
+    private Long deptId;
 
     /**
      * 用户头像
